cleaning up files/common functions
