+++ allennlp_models/vision/dataset_readers/vgqa.py
@@ -31,210 +31,10 @@ from allennlp.data.tokenizers import Tokenizer
 from allennlp.modules.vision.grid_embedder import GridEmbedder
 from allennlp.modules.vision.region_detector import RegionDetector
 
+from allennlp_models.vision.dataset_readers import utils
 from allennlp_models.vision.dataset_readers.vision_reader import VisionReader
 
 logger = logging.getLogger(__name__)
-
-# todo: put these in utils
-contractions = {
-    "aint": "ain't",
-    "arent": "aren't",
-    "cant": "can't",
-    "couldve": "could've",
-    "couldnt": "couldn't",
-    "couldn'tve": "couldn't've",
-    "couldnt've": "couldn't've",
-    "didnt": "didn't",
-    "doesnt": "doesn't",
-    "dont": "don't",
-    "hadnt": "hadn't",
-    "hadnt've": "hadn't've",
-    "hadn'tve": "hadn't've",
-    "hasnt": "hasn't",
-    "havent": "haven't",
-    "hed": "he'd",
-    "hed've": "he'd've",
-    "he'dve": "he'd've",
-    "hes": "he's",
-    "howd": "how'd",
-    "howll": "how'll",
-    "hows": "how's",
-    "Id've": "I'd've",
-    "I'dve": "I'd've",
-    "Im": "I'm",
-    "Ive": "I've",
-    "isnt": "isn't",
-    "itd": "it'd",
-    "itd've": "it'd've",
-    "it'dve": "it'd've",
-    "itll": "it'll",
-    "let's": "let's",
-    "maam": "ma'am",
-    "mightnt": "mightn't",
-    "mightnt've": "mightn't've",
-    "mightn'tve": "mightn't've",
-    "mightve": "might've",
-    "mustnt": "mustn't",
-    "mustve": "must've",
-    "neednt": "needn't",
-    "notve": "not've",
-    "oclock": "o'clock",
-    "oughtnt": "oughtn't",
-    "ow's'at": "'ow's'at",
-    "'ows'at": "'ow's'at",
-    "'ow'sat": "'ow's'at",
-    "shant": "shan't",
-    "shed've": "she'd've",
-    "she'dve": "she'd've",
-    "she's": "she's",
-    "shouldve": "should've",
-    "shouldnt": "shouldn't",
-    "shouldnt've": "shouldn't've",
-    "shouldn'tve": "shouldn't've",
-    "somebody'd": "somebodyd",
-    "somebodyd've": "somebody'd've",
-    "somebody'dve": "somebody'd've",
-    "somebodyll": "somebody'll",
-    "somebodys": "somebody's",
-    "someoned": "someone'd",
-    "someoned've": "someone'd've",
-    "someone'dve": "someone'd've",
-    "someonell": "someone'll",
-    "someones": "someone's",
-    "somethingd": "something'd",
-    "somethingd've": "something'd've",
-    "something'dve": "something'd've",
-    "somethingll": "something'll",
-    "thats": "that's",
-    "thered": "there'd",
-    "thered've": "there'd've",
-    "there'dve": "there'd've",
-    "therere": "there're",
-    "theres": "there's",
-    "theyd": "they'd",
-    "theyd've": "they'd've",
-    "they'dve": "they'd've",
-    "theyll": "they'll",
-    "theyre": "they're",
-    "theyve": "they've",
-    "twas": "'twas",
-    "wasnt": "wasn't",
-    "wed've": "we'd've",
-    "we'dve": "we'd've",
-    "weve": "we've",
-    "werent": "weren't",
-    "whatll": "what'll",
-    "whatre": "what're",
-    "whats": "what's",
-    "whatve": "what've",
-    "whens": "when's",
-    "whered": "where'd",
-    "wheres": "where's",
-    "whereve": "where've",
-    "whod": "who'd",
-    "whod've": "who'd've",
-    "who'dve": "who'd've",
-    "wholl": "who'll",
-    "whos": "who's",
-    "whove": "who've",
-    "whyll": "why'll",
-    "whyre": "why're",
-    "whys": "why's",
-    "wont": "won't",
-    "wouldve": "would've",
-    "wouldnt": "wouldn't",
-    "wouldnt've": "wouldn't've",
-    "wouldn'tve": "wouldn't've",
-    "yall": "y'all",
-    "yall'll": "y'all'll",
-    "y'allll": "y'all'll",
-    "yall'd've": "y'all'd've",
-    "y'alld've": "y'all'd've",
-    "y'all'dve": "y'all'd've",
-    "youd": "you'd",
-    "youd've": "you'd've",
-    "you'dve": "you'd've",
-    "youll": "you'll",
-    "youre": "you're",
-    "youve": "you've",
-}
-manual_map = {
-    "none": "0",
-    "zero": "0",
-    "one": "1",
-    "two": "2",
-    "three": "3",
-    "four": "4",
-    "five": "5",
-    "six": "6",
-    "seven": "7",
-    "eight": "8",
-    "nine": "9",
-    "ten": "10",
-}
-articles = ["a", "an", "the"]
-period_strip = re.compile(r"(?!<=\d)(\.)(?!\d)")
-comma_strip = re.compile(r"(\d)(\,)(\d)")
-punct = [
-    ";",
-    r"/",
-    "[",
-    "]",
-    '"',
-    "{",
-    "}",
-    "(",
-    ")",
-    "=",
-    "+",
-    "\\",
-    "_",
-    "-",
-    ">",
-    "<",
-    "@",
-    "`",
-    ",",
-    "?",
-    "!",
-]
-
-
-def process_punctuation(inText: str) -> str:
-    outText = inText
-    for p in punct:
-        if (p + " " in inText or " " + p in inText) or (comma_strip.search(inText) is not None):
-            outText = outText.replace(p, "")
-        else:
-            outText = outText.replace(p, " ")
-    outText = period_strip.sub("", outText, re.UNICODE)
-    return outText
-
-
-def process_digit_article(input: str) -> str:
-    output = []
-    for word in input.lower().split():
-        word = manual_map.get(word, word)
-        if word not in articles:
-            output.append(word)
-        else:
-            pass
-    for index, word in enumerate(output):
-        if word in contractions:
-            output[index] = contractions[word]
-    return " ".join(output)
-
-
-# todo: also remove words like a/the/an/etc?
-@lru_cache(maxsize=None)
-def preprocess_answer(answer: str) -> str:
-    answer = process_digit_article(process_punctuation(answer))
-    answer = answer.replace(",", "")
-    return answer
-
-
-def get_score(count: int) -> float:
-    return min(1.0, count / 3)
 
 
 @DatasetReader.register("vgqa")
@@ -273,10 +73,6 @@ class VGQAReader(VisionReader):
         returns.
     image_processing_batch_size: `int`
         The number of images to process at one time while featurizing. Default is 8.
-    multiple_answers_per_question: `bool`
-        VQA questions have multiple answers. By default, we use all of them, and give more
-        points to the more common answer. But VQA also has a special answer, the so-called
-        "multiple choice answer". If this is set to `False`, we only use that answer.
     """
 
     def __init__(
@@ -293,7 +89,6 @@ class VGQAReader(VisionReader):
         cuda_device: Optional[Union[int, torch.device]] = None,
         max_instances: Optional[int] = None,
         image_processing_batch_size: int = 8,
-        multiple_answers_per_question: bool = False,
         write_to_cache: bool = True,
     ) -> None:
         run_featurization = image_loader and image_featurizer and region_detector
@@ -329,7 +124,7 @@ class VGQAReader(VisionReader):
                 answer_vocab = cached_path(answer_vocab, extract_archive=True)
                 answer_vocab = Vocabulary.from_files(answer_vocab)
             self.answer_vocab = frozenset(
-                preprocess_answer(a)
+                utils.preprocess_answer(a)
                 for a in answer_vocab.get_token_to_index_vocabulary("answers").keys()
             )
 
@@ -346,8 +141,6 @@ class VGQAReader(VisionReader):
 
             if None in self.images:
                 del self.images[None]
-
-        self.multiple_answers_per_question = multiple_answers_per_question
 
     @overrides
     def _read(self, file_path: str):
@@ -404,7 +197,7 @@ class VGQAReader(VisionReader):
         attempted_instances_count = 0
         for qa, processed_image in zip(question_dicts, processed_images):
             question = qa["question"]
-            answer = preprocess_answer(qa["answer"])
+            answer = utils.preprocess_answer(qa["answer"])
             qa_id = qa["qa_id"]
 
             instance = self.text_to_instance(
